+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add logs while invoking `\llm`and `\\m+` command. [(#215)](https://github.com/dbcli/litecli/pull/215)
 * Support `--help` in the `\llm`and `\llm+` command. ([#214](https://github.com/dbcli/litecli/pull/214))
 * Make the history file location configurable. ([#206](https://github.com/dbcli/litecli/issues/206))
+* Add dot command to list views.
 
 ### Bug Fixes